Split code into different compilation units.
