+++ Instruments/BuildFile.xml
@@ -1,0 +1,4 @@
+<use name="boost"/>
+<use name="rootcore"/>
+<use name="AnalysisTools/Core"/>
+<export> <lib name="1"/> </export>
