+++ Print/BuildFile.xml
@@ -1,0 +1,5 @@
+<use name="boost"/>
+<use name="root"/>
+<use name="rootgpad"/>
+<use name="AnalysisTools/Core"/>
+<export> <lib name="1"/> </export>
