+++ cmake/include/common.cmake
@@ -23,7 +23,7 @@ get_filename_component(CMSSW_BASE_SRC "${AnalysisTools_DIR}" DIRECTORY)
 include_directories("${CMSSW_BASE_SRC}" "${AnalysisTools_DIR}/Core/include" "${PROJECT_SOURCE_DIR}")
 
 file(GLOB_RECURSE HEADER_LIST "*.h" "*.hh")
-file(GLOB_RECURSE SOURCE_LIST "*.cxx" "*.C" "*.cpp" "*.cc")
+file(GLOB_RECURSE SOURCE_LIST "*.cpp")
 file(GLOB_RECURSE EXE_SOURCE_LIST "*.cxx")
 file(GLOB_RECURSE SCRIPT_LIST "*.sh" "*.py")
 list(FILTER SCRIPT_LIST EXCLUDE REGEX "/__init__\\.py$")
@@ -63,8 +63,7 @@ set(TEST_LIST)
 foreach(exe_source ${EXE_SOURCE_LIST})
     get_filename_component(exe_name "${exe_source}" NAME_WE)
     message("Adding executable \"${exe_name}\"...")
-    add_executable("${exe_name}" "${exe_source}" "${RootDict}")
-    add_dependencies("${exe_name}" GenerateRootDict)
+    add_executable("${exe_name}" "${exe_source}")
     target_link_libraries("${exe_name}" ${ALL_LIBS})
     if(exe_name MATCHES "_t$")
         list(APPEND TEST_LIST "${exe_name}")
